refactor java.util.Date to java.time.LocalDate fixes #61
