+++ README.md
@@ -26,11 +26,11 @@ Jump over to the docs to get started.
 |----------------------|----------------------------|--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|------|
 | Virtual Customer     | Simulates orders           | [webflux](https://docs.spring.io/spring-framework/docs/current/reference/html/web-reactive.html), [jobrunr](https://github.com/jobrunr/jobrunr#readme), [faker](https://github.com/DiUS/java-faker#readme),                          | 8081 |
 | Order Service        | Place and manage orders    | [flyway](https://github.com/flyway/flyway#readme), eh-[kafka](https://spring.io/projects/spring-kafka), [jpa](https://spring.io/guides/gs/accessing-data-jpa/)                                                                       | 8082 |
-| Loyalty Service      | Manage reward points       | [redis](https://github.com/spring-projects/spring-data-redis/blob/main/src/main/asciidoc/reference/reactive-redis.adoc), eh-[kafka](https://spring.io/projects/spring-kafka), [jpa](https://spring.io/guides/gs/accessing-data-jpa/) | 8083 |
+| Loyalty Service      | Manage reward points       | [redis](https://github.com/spring-projects/spring-data-redis/blob/main/src/main/asciidoc/reference/reactive-redis.adoc), eh-[kafka](https://spring.io/projects/spring-kafka), [jpa](https://spring.io/guides/gs/accessing-data-jpa/) | 8703 |
 | Makeline Service     | Fulfil order queue         | [cosmodb](https://learn.microsoft.com/en-us/azure/cosmos-db/nosql/sdk-java-v4), eh-[kafka](https://spring.io/projects/spring-kafka), [jpa](https://spring.io/guides/gs/accessing-data-jpa/)                                          | 8084 |
 | Receipt generation   | Archive receipts           | blob storage                                                                                                                                                                                                                         | 8085 |
 | Virtual Worker       | Simulate order completion  | [cosmodb](https://learn.microsoft.com/en-us/azure/cosmos-db/nosql/sdk-java-v4)                                                                                                                                                       | 8086 |
-| Accounting Service   | Process and transform data | [cosmodb](https://learn.microsoft.com/en-us/azure/cosmos-db/nosql/sdk-java-v4), eh-[kafka](https://spring.io/projects/spring-kafka), [jpa](https://spring.io/guides/gs/accessing-data-jpa/)                                          | 8087 |
+| Accounting Service   | Process and transform data | [cosmodb](https://learn.microsoft.com/en-us/azure/cosmos-db/nosql/sdk-java-v4), eh-[kafka](https://spring.io/projects/spring-kafka), [jpa](https://spring.io/guides/gs/accessing-data-jpa/)                                          | 8707 |
 | nn-bindings          | GPT-3 responses to prompts | [az-openai](https://learn.microsoft.com/en-us/azure/cognitive-services/openai/quickstart?pivots=programming-language-python), python                                                                                                 | 8000 |
 | local eureka server  | Service discovery          | eureka                                                                                                                                                                                                                               | 8761 |
 | local config server  | Centralized Configurations | spring cloud config                                                                                                                                                                                                                  | 8088 |
