+++ ancillary/integration-tests/reddog-asa-local.postman_collection.json
@@ -142,12 +142,12 @@
 				"method": "GET",
 				"header": [],
 				"url": {
-					"raw": "http://localhost:8087/orders",
+					"raw": "http://localhost:8707/orders",
 					"protocol": "http",
 					"host": [
 						"localhost"
 					],
-					"port": "8087",
+					"port": "8707",
 					"path": [
 						"orders"
 					]
@@ -236,11 +236,11 @@
 					}
 				},
 				"url": {
-					"raw": "localhost:8083/orders",
+					"raw": "localhost:8703/orders",
 					"host": [
 						"localhost"
 					],
-					"port": "8083",
+					"port": "8703",
 					"path": [
 						"orders"
 					]
