+++ ancillary/local-config-service/src/main/resources/config/loyalty-service.yaml
@@ -40,7 +40,7 @@ management:
     health:
       show-details: ALWAYS
 server:
-  port: 8083
+  port: 8703
 springdoc:
   api-docs:
     path: /api-docs
