+++ loyalty-service/Dockerfile
@@ -22,7 +22,7 @@ RUN rpm -Uhv https://packages.microsoft.com/config/centos/7/packages-microsoft-p
     echo java -Xshare:dump && \
     java -Xshare:dump
 
-ARG PORT=8083
+ARG PORT=8703
 RUN mkdir /app
 COPY target/*.jar /app/app.jar
 WORKDIR /app
