+++ loyalty-service/Dockerfile-openjdk
@@ -7,5 +7,5 @@ COPY target/*.jar /app/app.jar
 WORKDIR /app
 RUN chown -R reddoguser:reddoguser /app
 USER reddoguser
-EXPOSE 8083
+EXPOSE 8703
 CMD "dumb-init" "java" "-jar" "app.jar"
