+++ loyalty-service/src/main/resources/application.yaml
@@ -44,7 +44,7 @@ management:
     health:
       show-details: ALWAYS
 server:
-  port: 8083
+  port: 8703
 springdoc:
   api-docs:
     path: /api-docs
