+++ order-service/pom.xml
@@ -180,7 +180,7 @@
             </plugin>
         </plugins>
     </build>
-<!--    <repositories>-->
+<!--    <repository>-->
 <!--        <repository>-->
 <!--            &lt;!&ndash;Configure and add personal access token in maven settings, see&ndash;&gt;-->
 <!--            &lt;!&ndash; https://devdiv.visualstudio.com/OnlineServices/_artifacts/feed/java-engineering-feed/connect/maven &ndash;&gt;-->
@@ -193,7 +193,7 @@
 <!--                <enabled>true</enabled>-->
 <!--            </snapshots>-->
 <!--        </repository>-->
-<!--    </repositories>-->
+<!--    </repository>-->
     <distributionManagement>
         <repository>
             <id>github</id>
